change in datatype of count

+++ src/main/java/org/littleshoot/udt/server/UdtFileUploadServer.java
@@ -32,7 +32,7 @@ import com.barchart.udt.util.LogUtil;
 
 public class UdtFileUploadServer {
 
-	private static int count;
+	private static long count;
 	private final Logger log = Logger.getLogger(getClass());
 	private final ExecutorService readPool = Executors.newCachedThreadPool();
